backend/speakspace_backend/requirements.txt
asgiref==3.8.1
channels==4.0.0
channels-redis==4.1.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
